Move bspline fit to separate module
